--- src/modules/plugins/ewq.js
@@ -1,1 +1,0 @@
-console.log("hello from st plugin!!!");
