+++ src/modules/plugins/modern-icons/modern-icons.css
@@ -1,5 +1,3 @@
-/* This file is loaded after scripts/doc-end.js is loaded */
-
 /* fallback */
 @font-face {
   font-family: "Material Icons Round";
