+++ src/modules/plugins/timetable-labels.js
@@ -1,4 +1,4 @@
-export default defineUnlistedScript(() => {
+export default defineWxtPlugin(() => {
   const days = ["Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday", "Sunday"];
 
   if (window.location.pathname === "/timetable" && document.getElementsByClassName("timetable")[0]) {
